yuridik/backend/requirements.txt
# Imtiyoz-AI Backend Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-google-genai>=0.0.6
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web Scraping
duckduckgo-search>=4.1.1
beautifulsoup4>=4.12.0
httpx>=0.26.0
lxml>=5.1.0

# Async & Utilities
asyncio-throttle>=1.0.2
sse-starlette>=1.8.2
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=23.12.0
isort>=5.13.0
